Update: refractor folder components
